introduce arduino.

+++ components/common/common_buffer.c
@@ -137,14 +137,14 @@ int buf_seek_abs(buffer_t *buf, uint32_t pos)
 {
     if (buf == NULL) return -1;
 
-    if(pos > buf->fill_pos) {
+    if(pos > (uint32_t)buf->fill_pos) {
         ESP_LOGE(TAG, "buf_seek_abs failed, pos = %u larger than fill_pos %u", pos, (uint32_t) buf->fill_pos);
         return -1;
     }
 
     size_t delta = pos - (uint32_t) buf->read_pos;
     buf->bytes_consumed += delta;
-    buf->read_pos = pos;
+    buf->read_pos = (uint8_t)pos;
 
     return 0;
 }
@@ -160,7 +160,7 @@ size_t buf_read(void * ptr, size_t size, size_t count, buffer_t *buf)
         return -1;
     }
 
-    uint8_t delay = 0;
+    uint16_t delay = 0;
     while(bytes_to_copy > buf_data_unread(buf) && delay < 5000) {
         fill_read_buffer(buf);
         vTaskDelay(50 / portTICK_PERIOD_MS);
